requirements.txt
requests
pillow
python-dotenv
# Dependências de Hardware (Instalar apenas no Raspberry Pi)
# RPi.GPIO
# adafruit-blinka
# adafruit-circuitpython-st7735
